python_service/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
torch>=2.3.0
transformers>=4.41.0
Pillow>=10.3.0
pydantic>=2.7.0
accelerate>=0.30.0
sentencepiece>=0.2.0
